Move changed suites to integration tests
