+++ test/integration/routes/site-assignment-checkout.test.ts
@@ -6,7 +6,7 @@ import { describeWithEnv } from "#test-utils/db.ts";
 import { createTestListing } from "#test-utils/db-helpers/listings.ts";
 import { mockFormRequest, mockRequest } from "#test-utils/mocks.ts";
 import { setupStripe } from "#test-utils/settings.ts";
-import { postExpectingNoCheckout } from "./_shared-checkout.ts";
+import { postExpectingNoCheckout } from "../../routes/_shared-checkout.ts";
 
 describeWithEnv(
   "routes > site assignment checkout validation",
