+++ test/integration/server/package-children.test.ts
