+++ test/integration/server/payments-success-refunds.test.ts
@@ -20,7 +20,7 @@ import { signMeta, singleItem } from "#test-utils/factories.ts";
 import { mockRequest } from "#test-utils/mocks.ts";
 // jscpd:ignore-end
 import { setupStripe } from "#test-utils/settings.ts";
-import { createHiddenPackageGroup } from "./payment-success-helpers.ts";
+import { createHiddenPackageGroup } from "../../lib/payment-success-helpers.ts";
 
 describeWithEnv("server (payment flow: ticket success)", { db: true }, () => {
   describe("GET /payment/success (ticket)", () => {
