+++ test/integration/server/payments/confirm.test.ts
