+++ test/integration/server/payments/success.test.ts
@@ -18,7 +18,7 @@ import {
   stubRefundPayment,
   stubRetrieveCheckoutSession,
 } from "#test-utils/webhooks.ts";
-import { fillSoldOutListing } from "./_shared-setup.ts";
+import { fillSoldOutListing } from "../../../lib/server-payments/_shared-setup.ts";
 
 // jscpd:ignore-end
