+++ test/integration/server/public/daily-listings-single.test.ts
@@ -14,7 +14,7 @@ import { describeWithEnv } from "#test-utils/db.ts";
 import { createTestHoliday } from "#test-utils/db-helpers/holidays.ts";
 import { mockFormRequest, mockRequest } from "#test-utils/mocks.ts";
 import { setupStripe } from "#test-utils/settings.ts";
-import { createDailyListing } from "./daily-listing.ts";
+import { createDailyListing } from "../../../lib/server-public/daily-listing.ts";
 
 // jscpd:ignore-end
