+++ test/integration/server/webhooks/can-pay-more-multi-ticket.test.ts
@@ -14,7 +14,7 @@ import {
   expectWebhookKeptAndRefunded,
   expectWebhookProcessed,
 } from "#test-utils/webhooks.ts";
-import { twoListingsAttendees } from "../attendee-read-helpers.ts";
+import { twoListingsAttendees } from "../../../lib/attendee-read-helpers.ts";
 
 // jscpd:ignore-end
