+++ test/integration/server/webhooks/modifiers.test.ts
@@ -20,7 +20,7 @@ import {
   checkoutSessionEvent,
   expectWebhookProcessed,
 } from "#test-utils/webhooks.ts";
-import { createServiceChargeScenario } from "./service-charge-scenario.ts";
+import { createServiceChargeScenario } from "../../../lib/server-webhooks/service-charge-scenario.ts";
 
 // jscpd:ignore-end
