+++ test/integration/server/webhooks/price-signature-stored-refund-and-ignore.test.ts
@@ -25,7 +25,7 @@ import {
   signedMeta,
   stubRefundOk,
   webhookRequest,
-} from "./helpers.ts";
+} from "../../../lib/webhook-price-signature/helpers.ts";
 
 describeWithEnv(
   "webhook signed price oracle — stored refunds & ignores",
