+++ test/integration/server/webhooks/registration-closed.test.ts
